% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sim_deviation.R
\name{compute_sim_deviation}
\alias{compute_sim_deviation}
\title{Compute the average simulated deviation for a fiven pair of individual}
\usage{
compute_sim_deviation(
  sim_files,
  results_df,
  sample_regex,
  label_request,
  percent = TRUE,
  simlabel = NULL,
  metric = NULL,
  subset = NULL,
  progressor = NULL
)
}
\description{
Compute the average simulated deviation for a fiven pair of individual
}
