% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bc_matrix.R
\name{get_bc_matrix}
\alias{get_bc_matrix}
\title{Generate a pair-specific matrix of Bhattacharya coefficients between every
pedigree comparisons distribution.}
\usage{
get_bc_matrix(sims_data, labels_to_keep)
}
\arguments{
\item{sims_data}{(dataframe) a pair-specific '.sims' file, containing
raw simulation results. This will most likely be the output of
grups.plots::load_simfile()}

\item{labels_to_keep}{(string) vector specifying which pedigree comparisons
labels should be kept within the matrix.}
}
\value{
a n*n matrix of Bhattacharya coeffients, where each column and
row corresponds to a specific pedigree comparison label.
}
\description{
Generate a pair-specific matrix of Bhattacharya coefficients between every
pedigree comparisons distribution.
}
