% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kinship_matrix.R
\name{get_kinship_matrix}
\alias{get_kinship_matrix}
\title{Generate a kinship matrix summarizing the classification results of the
'pedigree-sims' of GRUPS-rs.}
\usage{
get_kinship_matrix(data, sample_regex)
}
\arguments{
\item{data}{(dataframe) a GRUPS-rs '.result' dataframe, containing pairs of
individuals and their estimated most likely relationship. This will
most likely be the output of grups.plots::load_res_file()}

\item{sample_regex}{(string) Specify the regular expression used to extract
sample names from the 'Pair_name' column.}
}
\value{
A dataframe, representing a "melted" kinship matrix.
}
\description{
Generate a kinship matrix summarizing the classification results of the
'pedigree-sims' of GRUPS-rs.
}
