% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_norm_method.R
\name{get_norm_method}
\alias{get_norm_method}
\title{(private) Get the normalization value required to normalize the results of
the 'pwd-from-stin' module of GRUPS-rs}
\usage{
get_norm_method(norm_request, norm_method, norm_metric, norm_value = NULL)
}
\arguments{
\item{norm_request}{(bool) Whether or not results should be output as ratios
of the normalization value.}

\item{norm_method}{(string) User-defined string specifying which method, or
subset of comparisons should be used. Valid values are:
- "Pairwise": use pairwise comparisons
- "Self"    : use self comparisons
- "Value"   : use a predefined normalization value}

\item{norm_metric}{(string) User-defined normalization metric.
Valid values are:
- "Median" : use the median of all targeted comparisons
- "Mean"   : use the mean of all targeted comparisons
- "Min"    : use the minimum of all targeted comparisons
- "Max"    : use the maximum of all targeted comparisons.}

\item{norm_value}{(string) User input optional set normalization value.
This is only useful when setting norm_method = "Value"}
}
\value{
A lambda expression applying the appropriate summary statistic.
}
\description{
(private) Get the normalization value required to normalize the results of
the 'pwd-from-stin' module of GRUPS-rs
}
