% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_norm_metric.R
\name{get_norm_metric}
\alias{get_norm_metric}
\title{(private) Parse a user-defined string value into a base R function}
\usage{
get_norm_metric(norm_metric)
}
\arguments{
\item{norm_metric}{(string) User-defined normalization metric.
Valid values are:
- "Median" : use the median of all targeted comparisons
- "Mean"   : use the mean of all targeted comparisons
- "Min"    : use the minimum of all targeted comparisons
- "Max"    : use the maximum of all targeted comparisons.}
}
\value{
a summary statistic function corresponding to the provided string.
}
\description{
(private) Parse a user-defined string value into a base R function
}
