% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_norm_values.R
\name{get_norm_values}
\alias{get_norm_values}
\title{Normalize the results from the 'pwd-from-stdin' output of GRUPS-rs}
\usage{
get_norm_values(
  norm_method,
  norm_request = FALSE,
  norm_metric,
  norm_value,
  pwd_data
)
}
\arguments{
\item{norm_method}{(string) User-defined string specifying which method, or
subset of comparisons should be used. Valid values are:
- "Pairwise": use pairwise comparisons
- "Self"    : use self comparisons
- "Value"   : use a predefined normalization value}

\item{norm_request}{(bool) Whether or not results should be output as ratios
of the normalization value.}

\item{norm_metric}{(string) User-defined normalization metric.
Valid values are:
- "Median" : use the median of all targeted comparisons
- "Mean"   : use the mean of all targeted comparisons
- "Min"    : use the minimum of all targeted comparisons
- "Max"    : use the maximum of all targeted comparisons.}

\item{norm_value}{(sting) User input optional set normalization value.
This is only useful when setting norm_method = "Value"}

\item{pwd_data}{(dataframe) A \code{.pwd} results file dataframe. Most likely the
output of grups.plot::load_pairwise_file()}
}
\value{
a list of thresholds defining the degree of relatedness of each
pbserved pairwise misamtch rate.
}
\description{
Normalize the results from the 'pwd-from-stdin' output of GRUPS-rs
}
