% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_odds_matrix.R
\name{get_odds_matrix}
\alias{get_odds_matrix}
\title{Generate a pair-specific matrix of Odds Ratios between every
pedigree comparisons distribution.}
\usage{
get_odds_matrix(sims_data, observed_results, pair, labels_to_keep)
}
\arguments{
\item{sims_data}{(dataframe) a pair-specific '.sims' file, containing
raw simulation results. This will most likely be the output of
grups.plots::load_simfile()}

\item{observed_results}{(dataframe) a '.results' dataframe containing the
simulation results of the 'pedigree-sims' module of grups-rs. This
will most likely be the output of grups.plots::load_res_file().}

\item{pair}{(string) Specify which pairiwse comparison label should be
targeted to generate the matrix.}

\item{labels_to_keep}{(string) vector of labels specifying which pedigree
comparisons labels should be kept within the matrix.}
}
\value{
a n*n matrix of Log(Odds ratios), where each column and row
corresponds to a specific pedigree comparison label.
}
\description{
Generate a pair-specific matrix of Odds Ratios between every
pedigree comparisons distribution.
}
