% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sim_deviation.R
\name{get_sim_deviation_closest}
\alias{get_sim_deviation_closest}
\title{Compute the average simulated deviation between the observed PMR, and
the closest pedigree simulationdistribution.}
\usage{
get_sim_deviation_closest(sims_dataframe, obs_pwd, ...)
}
\arguments{
\item{sims_dataframe}{a pair-specific .sims dataframe.
Most likely the output of grups.plots::load_simfile()}

\item{obs_pwd}{a pair-specific observed pairwise mismatch rate}
}
\value{
a deviation ratio
}
\description{
Compute the average simulated deviation between the observed PMR, and
the closest pedigree simulationdistribution.
}
