% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hline.R
\name{hline}
\alias{hline}
\title{(private) Generate a simple horizontal line in plotly}
\usage{
hline(y, color = "red")
}
\arguments{
\item{y}{(numeric) vertical coordinate of the horizontal line}

\item{color}{(string) color of the horizontal line.}
}
\value{
(list) a list defining a simple horizontal line for plotly
}
\description{
(private) Generate a simple horizontal line in plotly
}
