% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_blockfile.R
\name{load_blockfile}
\alias{load_blockfile}
\title{Load a pair-specific '.blk' file into a simple dataframe.}
\usage{
load_blockfile(path, width, step = 1)
}
\arguments{
\item{path}{path leading to a GRUPS pair-specific '.blk' results file}

\item{width}{The desired which of every window (in Mbases)}

\item{step}{The desired sliding step between every window (in Mbases)}
}
\value{
a dataframe with columns "chr" "start" "end" "overlap" "pwd"
}
\description{
Load a pair-specific '.blk' file into a simple dataframe.
}
