% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pairwise_file.R
\name{load_pairwise_file}
\alias{load_pairwise_file}
\title{Import and merge the output '.pwd' and '.result' fils of GRUPS-rs into a
single dataframe.}
\usage{
load_pairwise_file(
  path,
  res_data,
  norm_avg_type,
  sample_regex,
  min_overlap = 0,
  norm_request = FALSE,
  norm_method = "Raw",
  norm_metric = "Pairwise-comparisons",
  norm_value = NULL
)
}
\arguments{
\item{path}{(string) path leading to a GRUPS-rs '.pwd' file.}

\item{res_data}{(string) path leading to a GRUPS-rs '.result' file}

\item{norm_avg_type}{(string) Specify whether to prioritize the corrected
results of the pedigree-sims module, or the raw results of the
'pwd-from-stdin' module. Valid values are either "Corr." or "Raw."}

\item{sample_regex}{Specify the PCRE regular expression to use to parse
names from the Pair_name column.}

\item{min_overlap}{Specify a minimum overlap threshold to filter out
comparisons below this value.}

\item{norm_metric}{(string) User-defined string specifying which method, or
subset of comparisons should be used. Valid values are:
- "Pairwise": use pairwise comparisons
- "Self"    : use self comparisons
- "Value"   : use a predefined normalization value}

\item{norm_value}{(sting) User input optional set normalization value.
This is only useful when setting norm_method = "Value"}
}
\value{
a dataframe containing merged, normalized results.
}
\description{
Import and merge the output '.pwd' and '.result' fils of GRUPS-rs into a
single dataframe.
}
