% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_res_file.R
\name{load_res_file}
\alias{load_res_file}
\title{Utility wrapper to load a '.result' file into a dataframe}
\usage{
load_res_file(path)
}
\arguments{
\item{path}{path leading to a GRUPS-rs .result summary file}
}
\value{
dataframe containing simulation results for all pairs.
}
\description{
Utility wrapper to load a '.result' file into a dataframe
}
