% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_simfile.R
\name{load_simfile}
\alias{load_simfile}
\title{Utility wrapper to load a pair-specific '.sims' file into a dataframe}
\usage{
load_simfile(path)
}
\arguments{
\item{path}{path leading to a GRUPS '.sims' results file}
}
\value{
dataframe containing simulation results for a given pair.
}
\description{
Utility wrapper to load a pair-specific '.sims' file into a dataframe
}
