% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_svmop_probs.R
\name{load_svmop_probs}
\alias{load_svmop_probs}
\title{Utility wrapper to load a '.probs' file into a dataframe.}
\usage{
load_svmop_probs(path)
}
\arguments{
\item{path}{path leading to a GRUPS '.probs' summary file}
}
\value{
dataframe containing svmop_probs for all pairs.
}
\description{
Utility wrapper to load a '.probs' file into a dataframe.
}
