% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_pwd_results.R
\name{merge_pwd_results}
\alias{merge_pwd_results}
\title{(private) Merge a '.pwd' dataframe with a '.results' dataframe}
\usage{
merge_pwd_results(pwd_df = NULL, res_df)
}
\arguments{
\item{pwd_df}{a dataframe of raw PWD results (.pwd    file extension)}

\item{res_df}{a dataframe of pedsims results (.result file extension)}
}
\value{
either pwd_df or a merged dataframe of both if res_df is defined
(merge across pair names).
}
\description{
(private) Merge a '.pwd' dataframe with a '.results' dataframe
}
