% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_annotation.R
\name{ms_annotation}
\alias{ms_annotation}
\title{(private) Simple plotly annotation to plot the MS values
within the Observed pairwise mismatch rate plot.}
\usage{
ms_annotation(label, y)
}
\arguments{
\item{label}{label of the annotation}

\item{y}{coordinate of the text annotation}
}
\value{
plotly list annotation
}
\description{
(private) Simple plotly annotation to plot the MS values
within the Observed pairwise mismatch rate plot.
}
