% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bc_matrix.R
\name{plot_bc_matrix}
\alias{plot_bc_matrix}
\title{Plot the results of 'grups.plots::get_bc_matrix()'}
\usage{
plot_bc_matrix(
  bc_matrix,
  pair,
  plot_title = NULL,
  marker_text = NULL,
  cutoff_values,
  cutoff_labels,
  cutoff_colors = c("#44AA99", "#DDCC77", "#CC6677"),
  right_align = FALSE,
  absolute_values = FALSE
)
}
\arguments{
\item{pair}{(string) Label of the targeted pairwise comparison}

\item{plot_title}{(string) Title of the plot}

\item{marker_text}{(string) Add a prefix text to every cell within the
plotted matrix}

\item{cutoff_values}{\link{numeric} a vector of values used to define
significance cutoffs.}

\item{cutoff_colors}{\link{string} a vector of colors used to define
every cutoff value}

\item{right_align}{(bool) align cutoff values to the right}

\item{absolute_values}{(bool) plot raw odds ratios as absolute values.}

\item{matrix}{(matrix) matrix of bhattacharya coefs between sim
distributions.}

\item{cutoff}{labels \link{string} a vector of strings used to define
each cutoff value}
}
\value{
a plotly plot
}
\description{
Plot the results of 'grups.plots::get_bc_matrix()'
}
