% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_kinship_matrix.R
\name{plot_kinship_matrix}
\alias{plot_kinship_matrix}
\title{Plot the output of grups.plots::get_kinship_matrix() using the catmaply
package.}
\usage{
plot_kinship_matrix(
  kinship_matrix,
  dimensions,
  order,
  color_palette = NULL,
  ...
)
}
\arguments{
\item{kinship_matrix}{Kinship matrix}

\item{dimensions}{Shiny input dimensions. in pixels c(row, height)}

\item{order}{Order in which kinship degrees should be arranged.}

\item{color_palette}{specify a custom color palette for every kinship order.}

\item{...}{Additional catmaply::catmaply() arguments.}
}
\value{
a plotly kinship matrix
}
\description{
Plot the output of grups.plots::get_kinship_matrix() using the catmaply
package.
}
