% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_or_confidence.R
\name{plot_or_confidence}
\alias{plot_or_confidence}
\title{Plot the output of grups.plots::get_odds_matrix() as a scatter plot,
highlighting the odds of the most likely relationship vs. other odds}
\usage{
plot_or_confidence(or_matrix, predictor = NULL)
}
\arguments{
\item{or_matrix}{matrix of odds_ratio between distributions.}

\item{predictor}{label of the most likely relationship}
}
\value{
a plotly scatter plot
}
\description{
Plot the output of grups.plots::get_odds_matrix() as a scatter plot,
highlighting the odds of the most likely relationship vs. other odds
}
