% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pairwise_diff.R
\name{plot_pairwise_diff}
\alias{plot_pairwise_diff}
\title{Generate a barplot of all the observed pairwise differences.}
\usage{
plot_pairwise_diff(
  data,
  hide_self_comparisons = FALSE,
  norm_method = NULL,
  norm_metric = NULL,
  norm_values = NULL
)
}
\arguments{
\item{hide_self_comparisons}{(bool) whether to hide self-comparisons
from the plot}

\item{norm_metric}{(string) User-defined normalization metric.
Valid values are:
- "Median" : use the median of all targeted comparisons
- "Mean"   : use the mean of all targeted comparisons
- "Min"    : use the minimum of all targeted comparisons
- "Max"    : use the maximum of all targeted comparisons}

\item{plot_data}{a dataframe generated by grups.plots::load_pairwise_file()}
}
\value{
plotly barplot
}
\description{
Generate a barplot of all the observed pairwise differences.
}
