% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pedigree_sims.R
\name{plot_pedigree_sims}
\alias{plot_pedigree_sims}
\title{Plot pair-specific pedigree simulation results, in the form of a set
of violin plots vs. the observed pairwise mismatch rate of that pair.}
\usage{
plot_pedigree_sims(sims_dataframe, results_data, labels_to_plot, pair)
}
\arguments{
\item{sims_dataframe}{(dataframe) a dataframe of pair-specific '.sims'
raw simulations results}

\item{results_data}{(dataframe) a dataframe generated by the function
grups.plots::load_pairwise_file()}

\item{labels_to_plot}{(list) a list of pedigree comparison labels to
plot.}

\item{pair}{label of the target pairwise comparison to plot}
}
\value{
A plotly violin plot
}
\description{
Plot pair-specific pedigree simulation results, in the form of a set
of violin plots vs. the observed pairwise mismatch rate of that pair.
}
