% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sliding_windows.R
\name{plot_sliding_window}
\alias{plot_sliding_window}
\title{Generate a scatter plot of per-chromosome pairwise mismatch rates
within sliding windows, for a specific pairwise comparison.}
\usage{
plot_sliding_window(block_dataframe, pair)
}
\arguments{
\item{block_dataframe}{dataframe of pair-specific '.blk' file. Use the
output of grups.plots::load_blockfile()}

\item{pair}{label of the targeted pairwise comparison.}
}
\value{
A plotly scatterplot of per-chromosome windowed mismatch rates.
}
\description{
Generate a scatter plot of per-chromosome pairwise mismatch rates
within sliding windows, for a specific pairwise comparison.
}
