% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_reactive_checkbox.R
\name{shiny_reactive_checkbox}
\alias{shiny_reactive_checkbox}
\title{(private) generate a reactive Shiny UI checkbox}
\usage{
shiny_reactive_checkbox(values, title, input_id, ncol = 2)
}
\arguments{
\item{values}{\link{string} vector of checkbox elements.}

\item{title}{(string) title of the div / section.}

\item{input_id}{shiny input id of the target shiny html/css.}

\item{ncol}{number of display columns.}
}
\value{
list containing shiny css style, header, and checkboxgroupinput.
}
\description{
(private) generate a reactive Shiny UI checkbox
}
