% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_normality.R
\name{test_normality}
\alias{test_normality}
\title{Apply a Kolmogorov-Smirnov test of normality on every simulated
distribution of pedigree comparisons, for a given pair of indivduals.}
\usage{
test_normality(sims_file, alpha = 0.05)
}
\arguments{
\item{sims_file}{a pair-specific '.sims' dataframe. Most likely the output
of grups.plots::load_simfile()}

\item{alpha}{alpha risk treshold required to reject H0}
}
\description{
Apply a Kolmogorov-Smirnov test of normality on every simulated
distribution of pedigree comparisons, for a given pair of indivduals.
}
