% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltips.R
\name{tooltips}
\alias{tooltips}
\title{(private) Wrapper which outputs a dataframe containing tooltips
for every Shiny widget.}
\usage{
tooltips()
}
\value{
a dataframe of one row, where each column corresponds to
a static shiny inputId, and values correspond to the text
contents of a tooltip
}
\description{
(private) Wrapper which outputs a dataframe containing tooltips
for every Shiny widget.
}
